% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpiu.R
\name{add_wrss}
\alias{add_wrss}
\title{Calculate the WRSS}
\usage{
add_wrss(df_test, weights_by_status, lambda_pred, length_cpius, time_point)
}
\arguments{
\item{df_test}{a data frame contains composite event observations}

\item{weights_by_status}{a vector, the weights assign to each event type}

\item{lambda_pred}{a matrix contains the predicated hazard rate for each subject at each interval}

\item{length_cpius}{a vector, the length of each CPIU}

\item{time_point}{a scalar, the time point to evaluate WRSS}
}
\value{
a data frame contains WRSS for each subject
}
\description{
Calculate the WRSS
}
\details{
Calculate the WRSS(Gerds T. 2006) to evaluate the performance of \code{lambda_pred} when the true Y is unknown
}
