% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpiu.R
\name{add_wt}
\alias{add_wt}
\title{convert the recorded event time per patient to the number of events and wt at the given time point per patient}
\usage{
add_wt(data_to_convert, weights_by_status, time_point)
}
\arguments{
\item{data_to_convert:}{a data frame with columns of Id, X, Status}

\item{weights_by_status:}{a vector of weights by status, default is c(0,1,1) for censoring (status: 0), terminal event(status: 1) and recurrent event(status: 1)}

\item{time:}{the time point to calculate the number of events and wt}
}
\value{
a dataframe with number of events and wt at the given time point per patient
}
\description{
convert the recorded event time per patient to the number of events and wt at the given time point per patient
}
