% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpiu.R
\name{add_Y_hat}
\alias{add_Y_hat}
\title{Calculate the predicted number of events at given time points}
\usage{
add_Y_hat(lambda_pred, length_cpius, time_to_evaluate)
}
\arguments{
\item{lambda_pred}{a matrix of predicted hazard rates at each interval for multiple subjects}

\item{time_to_evaluate}{a scalar, time point to evaluate \code{Y^\hat}}

\item{interval_cpius}{a vector of lengths for each CPIU, same length with \code{ncol(lambda_pred)}}
}
\value{
a matrix of predicted number of events at each \code{time_points} for multiple subjects
}
\description{
Calculate the predicted number of events at given time points
}
\examples{
lambdas <- matrix(1:10, nrow = 2, byrow = TRUE)
intervals <- rep(3,5)
add_Y_hat(lambdas, intervals, c(3, 6, 9))
}
