% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpiu.R
\name{break_length_by_interval}
\alias{break_length_by_interval}
\title{break the given length into several intervals with the given intervals}
\usage{
break_length_by_interval(length, interval)
}
\arguments{
\item{length:}{the length to be broken}

\item{interval:}{the given intervals}
}
\value{
: a vector of length of each interval
}
\description{
break the given length into several intervals with the given intervals
}
\examples{
# example code
 break_length_by_interval(100, 20:26)
}
