% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compo_sim.R
\name{compo_sim}
\alias{compo_sim}
\title{composite event generator for simulation}
\usage{
compo_sim(
  n_patients = 1000,
  n_vars = 10,
  vars_cate = c(rep("binary", 6), rep("continuous", 4)),
  true_beta = c(0, 0, 0, 0.6, 0, 0, 0.8, 0, 0.7, 0),
  lambda = 0.01,
  constant_baseline_hazard = TRUE,
  a_shape_weibull = 4,
  sigma_scale_weibull = 4,
  sigma_scale_gamma = 0.05,
  non_linear_hazard = FALSE,
  non_linear_function = NULL,
  seed = 926,
  verbose = FALSE
)
}
\arguments{
\item{n_patients}{number of patients}

\item{n_vars}{number of covariates}

\item{vars_cate}{vector of "continuous", "binary"}

\item{true_beta}{True parameters}

\item{lambda}{baseline death hazard for entire population}

\item{constant_baseline_hazard}{whether to use constant baseline hazard; default is TRUE;
otherwise, use Weibull baseline hazard with shape parameter a_shape_weibull
and scale parameter sigma_scale_weibull}

\item{a_shape_weibull}{shape parameter of Weibull distribution}

\item{sigma_scale_weibull}{scale parameter of Weibull distribution}

\item{sigma_scale_gamma}{scale parameter of gamma distribution}

\item{non_linear_hazard}{whether to use non-linear hazard function}

\item{non_linear_function}{the non-linear function to generate hazard}

\item{seed}{seed for random number generation}

\item{verbose}{whether to print out summary of the number of recurrent events per patients}
}
\value{
a list of simulated data and parameters
}
\description{
composite event generator for simulation by non-homogenuous Poisson process
}
