% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpiu.R
\name{counts_by_interval_and_id}
\alias{counts_by_interval_and_id}
\title{calculate the number of events in each interval}
\usage{
counts_by_interval_and_id(
  event_times,
  ids,
  status,
  weights_by_status,
  interval
)
}
\arguments{
\item{event_times:}{a vector of event times}

\item{ids:}{a vector of ids}

\item{status:}{a vector of status}

\item{weights_by_status:}{a vector of weights by status}

\item{interval:}{a vector of intervals}
}
\value{
a matrix of number of events in each interval by id, row is id, column is interval
}
\description{
calculate the number of events in each interval
}
