% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_dummy.R
\name{cpius_to_dummy}
\alias{cpius_to_dummy}
\title{Dummy-encode factor/character covariates in CPIU object}
\usage{
cpius_to_dummy(object, cols_to_keep = NULL)
}
\arguments{
\item{object}{A CPIU object.}

\item{cols_to_keep}{Optional vector of column names to keep and consider for dummy encoding. If NULL, all columns will be considered for dummy-encoded.}
}
\value{
A CPIU object with dummy-encoded covariates in the \code{designMatrixY}.
}
\description{
This function takes a CPIU object and dummy-encodes any factor or character covariates
in the design matrix for the outcome variable Y. It removes the reference level
for each factor to avoid multicollinearity.
}
