% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_cpiu.R
\name{empirical_Y}
\alias{empirical_Y}
\title{Empirical Mean Number of Events at the Different Time Point}
\usage{
empirical_Y(
  compo_sim_list,
  weibull_baseline = TRUE,
  x,
  time_points,
  n_sims,
  n_size
)
}
\arguments{
\item{compo_sim_list}{simulation list either from \code{compo_sim} or \code{compo_sim_mao}}

\item{weibull_baseline}{which simulation scheme is used, default is \code{compo_sim}}

\item{x}{patient characteristics to exam}

\item{time_points}{time point to exam}

\item{n_sims}{number of simulations to run}

\item{n_size}{number of patients at each simulation}
}
\description{
Empirical Mean Number of Events at the Different Time Point
}
\details{
Calculate the empirical mean number of event at the different
time points by using data simulation
}
