% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpiu.R
\name{km_fit}
\alias{km_fit}
\title{fit kaplan meier survival function}
\usage{
km_fit(time, status)
}
\arguments{
\item{time:}{time to event}

\item{status:}{event status, 0 is censored, 1 is terminal event}
}
\value{
a list of unique time and survival probability
}
\description{
fit kaplan meier survival function
}
