% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{loadRforce}
\alias{loadRforce}
\title{loadRforce function for Rforce}
\usage{
loadRforce(path)
}
\arguments{
\item{path}{A character string giving the name of the path from which the object should be loaded}
}
\description{
loadRforce function for Rforce
}
