% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor.R
\name{manual_censoring}
\alias{manual_censoring}
\title{censoring the given dataset manually}
\usage{
manual_censoring(data_to_convert, censoring_quantile)
}
\arguments{
\item{data_to_convert}{the dataset to be converted; including the columns of \code{Id}, \code{Status} and \code{Time}}

\item{censoring_quantile}{the quantile of observed time to be censored}
}
\value{
the censored dataset
}
\description{
censoring the given dataset based on the given quantile of observed time
}
\examples{
list_data_to_convert <- compo_sim()
df_converted <- manual_censoring(
    list_data_to_convert$dataset,
    0.9
)
str(df_converted)

}
