% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_cpiu.R
\name{observed_risk_time}
\alias{observed_risk_time}
\title{calculate observed risk time at the given interval}
\usage{
observed_risk_time(Gt, X, Status = 0, low = 1, high = 3)
}
\arguments{
\item{Gt:}{a list of unique time and survival probability of Censoring}

\item{X:}{the time point associated with the Status}

\item{Status:}{the event status at the time point X}

\item{low:}{the lower bound of the interval}

\item{high:}{the upper bound of the interval}
}
\value{
: the observed risk time at the given interval
}
\description{
calculate observed risk time at the given interval
}
