% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_cpiu.R
\name{patients_to_cpius}
\alias{patients_to_cpius}
\title{convert the recorded event time per patient to the number of events per interval per patient}
\usage{
patients_to_cpius(
  data_to_convert,
  units_of_cpiu,
  weights_by_status = c(0, 1, 1),
  pseudo_risk = TRUE,
  wide_format = TRUE
)
}
\arguments{
\item{data_to_convert:}{a data frame with columns of Id, X, Status}

\item{units_of_cpiu:}{a vector of units of CPIU}

\item{weights_by_status:}{a vector of weights by status, default is c(0,1,1) for censoring (status: 0), terminal event(status: 1) and recurrent event(status: 1)}

\item{pseudo_risk:}{a boolean value indicating whether to use pseudo risk time}
}
\value{
: a dataframe with number of events in each interval by id, row is id, column is interval
}
\description{
convert the recorded event time per patient to the number of events per interval per patient
}
