% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{predict.Rforce}
\alias{predict.Rforce}
\title{Predict function for Rforce}
\usage{
\method{predict}{Rforce}(forest, designMatrix)
}
\arguments{
\item{forest}{An external pointer to the forest object created by Rforce}

\item{designMatrix}{A matrix of design features for prediction}
}
\value{
A list containing the predicted values
}
\description{
Predict function for Rforce
}
