% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpiu.R
\name{pseudo_risk_time}
\alias{pseudo_risk_time}
\title{calculate pseudo risk time at the given interval}
\usage{
pseudo_risk_time(Gt, X, Status = 0, low = 1, high = 3)
}
\arguments{
\item{Gt:}{a list of unique time and survival probability of Censoring}

\item{X:}{the time point associated with the Status}

\item{Status:}{the event status at the time point X}

\item{low:}{the lower bound of the interval}

\item{high:}{the upper bound of the interval}
}
\value{
: the pseudo risk time at the given interval
}
\description{
calculate pseudo risk time at the given interval
}
\examples{
# example code
 Gt <- list(unique_time = c(0, 1, 1.8, 3.5, 4),
            surv_prop = c(1, 0.8,0.6, 0.3, 0.1))
pseudo_risk_time(Gt, X =1.5 , Status = 1, low =0.5, high=1.9)
}
