% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{RC_ColsPermute}
\alias{RC_ColsPermute}
\title{ColsPermute}
\usage{
RC_ColsPermute(x, colsToPermute, seed)
}
\arguments{
\item{x}{A matrix, it is stored in column-major order when passed to C}
}
\description{
ColsPermute
}
