% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{Rforce}
\alias{Rforce}
\title{Main Rforce function}
\usage{
Rforce(
  design_matrix_Y,
  auxiliary_features,
  variable_Ids,
  units_of_cpius,
  split_rule = c("Rforce-QLR", "Rforce-GEE", "Rforce-GEEIntr", "RF-SLAM"),
  n_trees = 500,
  max_depth = 8,
  min_node_size = 10,
  min_gain = 0,
  mtry = NA,
  n_splits = NA,
  seed = 926
)
}
\arguments{
\item{design_matrix_Y}{data.frame that includes design matrix and response(number of events at each CPIU)}

\item{auxiliary_features}{data.frame that includes Id, X, Status, pseudo-risktime-at-duration}
}
\description{
Main Rforce function
}
