% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_sim.R
\name{surv_sim}
\alias{surv_sim}
\title{Survival data generation using either weibull or inverse CDF method}
\usage{
surv_sim(
  n_patients,
  n_vars,
  vars_cate,
  scale_weibull,
  shape_weibull,
  beta,
  rateC,
  seed = 926,
  method = c("InverseCDF", "Weibull")
)
}
\arguments{
\item{n_patients}{number of patients}

\item{n_vars}{number of covariates}

\item{vars_cate}{vector of "continuous", "binary"}

\item{scale_weibull}{scale parameter for baseline hazard of weibull distribution}

\item{shape_weibull}{shape parameter of weibull, shape_weibull < 1 decreasing hazard, shape_weibull > 1 increasing hazard, shape_weibull=1 constant hazard}

\item{beta}{Fixed effect parameter, could control noise variable by set beta\link{i} = 0}

\item{rateC}{censoring rate,}

\item{seed}{seed}

\item{method}{"InverseCDF" or "Weibull"}
}
\value{
data frame with columns: Id, HazardWOBaseline, Time, Censor, X, Status, x1, x2, ...
}
\description{
Survival data generation using either weibull or inverse CDF method
}
\examples{
set.seed(926)
nsim <- 100
beta_true <- c(0, 0.8, -0.8, -1.5, 0, 3, 0, -1, 0.8, 1.5, 3, 0, -0.8, 0, -3)
betaHat <- matrix(nrow = nsim, ncol = 15)
for (k in 1:nsim) {
  df_surv <- surv_sim(
    n_patients = 1000,
    scale_weibull = 1,
    shape_weibull = 1,
    beta = beta_true,
    rateC = 0.3,
    n_vars = 15,
    vars_cate = c(rep("binary", 7), rep("continous", 8)),
    method = "InverseCDF",
    seed = 926 + k
  )
  fit <- coxph(Surv(X, Status) ~ ., data = df_surv[, -c(1:4)])
  betaHat[k, ] <- fit$coef
}
colMeans(betaHat) - beta_true
}
