% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpiu.R
\name{trueY}
\alias{trueY}
\title{Calculate the mean number events for each subject at different time points in the simulated dataset}
\usage{
trueY(compo_sim_list, time_points)
}
\arguments{
\item{compo_sim_list}{the object from the \code{compo_sim} or \code{compo_sim_mao}}

\item{time_points}{the time points that need to be evaluated}
}
\value{
a data frame (n_subject x length(time_point))
}
\description{
Calculate the mean number events for each subject at different time points in the simulated dataset
}
