% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_cpiu.R
\name{validate.CPIU}
\alias{validate.CPIU}
\title{validate CPIU object}
\usage{
\method{validate}{CPIU}(object)
}
\arguments{
\item{object:}{a CPIU object}
}
\value{
: TRUE if the object is valid
}
\description{
validate CPIU object
}
