% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_cpiu.R
\name{validate.data.frame}
\alias{validate.data.frame}
\title{validate data.frame object}
\usage{
\method{validate}{data.frame}(object, required_cols = NULL)
}
\arguments{
\item{object:}{a data.frame object}

\item{required_cols:}{a vector of required column names}
}
\value{
: TRUE if the object is valid
}
\description{
validate data.frame object
}
