% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vimp.Rforce}
\alias{vimp.Rforce}
\title{Variable importance function for Rforce object}
\usage{
vimp.Rforce(forest)
}
\arguments{
\item{object}{An Rforce object}
}
\value{
A matrix containing variable importance statistics for each variable in each tree of the forest,
row correspond to trees and columns correspond to variables.
}
\description{
Variable importance function for Rforce object
}
