% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcompo_kim.R
\name{wcompo_est}
\alias{wcompo_est}
\title{Fast implementation of Lu Mao's Wcompo method by Kim So Young}
\usage{
wcompo_est(data, weight)
}
\arguments{
\item{data}{data frame including Id, X, Status, and covariates}

\item{weight}{weight for death and recurrent events}
}
\value{
list, including beta estimate, y, and t
}
\description{
This function is a fast implementation of Lu Mao's Wcompo method by Kim So Young
}
\examples{
data_list <- compo_sim(
  n_patients = 500,
  constant_baseline_hazard = FALSE,
  seed = 926,
  verbose = FALSE
)
library(dplyr)
df_train <- data_list[[1]] \%>\%
  dplyr::mutate(X = Time) \%>\%
  dplyr::select(-c("Time"))
estimate_list <- wcompo_est(
  data = df_train,
  weight = c(1, 1)
)
estimate_list$beta
}
