% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpiu.R
\name{Y_hat_by_time}
\alias{Y_hat_by_time}
\title{Calculate the predicted number of events at given time points}
\usage{
Y_hat_by_time(lambdas, interval_lengths, time_points)
}
\arguments{
\item{lambdas}{a vector of predicted hazard rates at each interval}

\item{interval_lengths}{a vector of lengths for each interval, same length with \code{lambdas}}

\item{time_points}{a vector of time points that been evaluated}
}
\value{
a vector of predicted number of events for each \code{time_points}
}
\description{
Calculate the predicted number of events at given time points
}
\examples{
lambdas <- 1:10
intervals <- rep(3, 10)
Y_hat_by_time(lambdas, intervals, c(3, 6, 9))
}
