% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpiu.R
\name{add_wt}
\alias{add_wt}
\title{convert the recorded event time per patient to the number of events and wt at the given time point per patient}
\usage{
add_wt(data_to_convert, weights_by_status, time_to_evaluate)
}
\arguments{
\item{time_to_evaluate}{a scalar, or vector with length equal to \code{length(unique(data_to_convert$Id))}
the time point to calculate the number of events and wt for each patient;}

\item{data_to_convert:}{a data frame with columns of Id, X, Status}

\item{weights_by_status:}{a vector of weights by status, default is \code{c(0,1,1)} for censoring (status: 0), terminal event(status: 1) and recurrent event(status: 1)}
}
\value{
a data.frame with number of events and wt at the given time point per patient
}
\description{
convert the recorded event time per patient to the number of events and wt at the given time point per patient
}
