% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{Rforce}
\alias{Rforce}
\title{Main Rforce function}
\usage{
Rforce(
  data = NULL,
  formula = NULL,
  id = "Id",
  n_intervals = 10,
  design_matrix_Y = NULL,
  auxiliary_features = NULL,
  variable_Ids = NULL,
  units_of_cpius = NULL,
  event_types = NULL,
  event_weights = NULL,
  split_rule = c("Rforce-QLR", "Rforce-GEE", "Rforce-GEEIntr", "RF-SLAM"),
  n_trees = NULL,
  max_depth = NULL,
  min_node_size = NULL,
  min_gain = NULL,
  mtry = NA,
  n_splits = NA,
  seed = 926
)
}
\arguments{
\item{design_matrix_Y}{data.frame that includes design matrix and response(number of events at each CPIU)}

\item{auxiliary_features}{data.frame that includes Id, X, Status, pseudo-risktime-at-duration}
}
\description{
Main Rforce function
}
