% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{saveRforce}
\alias{saveRforce}
\title{saveRDS function for Rforce}
\usage{
saveRforce(forest, file, ...)
}
\arguments{
\item{file}{A character string giving the name of the file where the object should be saved}

\item{object}{An Rforce object}
}
\description{
saveRDS function for Rforce
}
